requirements.txt
Django==5.1.1
django-import-export==4.2.0
django-jazzmin @ git+https://github.com/thercior/django-jazzmin.git@15200d96ba7f6177bcf6c3ec4d09e43ed6ca0325
djangorestframework==3.15.2
psycopg==3.2.2
psycopg-binary==3.2.2
python-environ==0.4.54
requests==2.32.3